requirements.txt
# Core MCP
fastmcp>=2.0.0
mcp>=1.0.0

# LangChain / LangGraph
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# Database
psycopg[binary]>=3.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# HTTP Server (for FastMCP HTTP transport)
uvicorn>=0.30.0
httpx>=0.27.0

# Web UI
streamlit>=1.40.0
